+++ app/src/assets/themes/index.d.ts
@@ -1,16 +1,9 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Theme } from '@mui/material';
 import '@mui/material/ButtonBase';
 import '@mui/material/InputBase';
 import '@mui/material/styles/createTheme';
-
-declare module '@mui/styles/defaultTheme' {
-  interface DefaultTheme extends Theme {
-    noEmptyExtends: boolean;
-  }
-}
 
 declare module '@mui/material/InputBase' {
   interface InputBaseProps {