+++ app/src/components/DurationFieldWrapper/DurationFieldWrapper.tsx
@@ -1,14 +1,11 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Box, BoxTypeMap, Typography } from '@mui/material';
-import { DefaultComponentProps } from '@mui/material/OverridableComponent';
+import { Box, BoxProps, Typography } from '@mui/material';
 import { PropsWithChildren } from 'react';
 import { useTranslation } from 'react-i18next';
 
-export const DurationFieldWrapper = (
-  props: PropsWithChildren<DefaultComponentProps<BoxTypeMap<Record<string, unknown>, 'div'>>>
-) => {
+export const DurationFieldWrapper = (props: PropsWithChildren<BoxProps>) => {
   const { t } = useTranslation();
 
   return (