+++ app/src/components/ui/SelectableList/index.tsx
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
-//
-// SPDX-License-Identifier: EUPL-1.2
-import SelectableList from './SelectableList';
-
-export * from './SelectableListItem';
-export default SelectableList;